feat : group router 수정 및 파일 정리
