+++ http/app.http
@@ -1,0 +1,8 @@
+###조회
+GET http://localhost:3000/api/image
+Content-Type: multipart/form-data
+
+{
+	"image": "File"
+}
+
